requirements.txt
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
PyYAML==6.0.1
tqdm==4.66.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0
